+++ NEWS.md
@@ -1,3 +1,13 @@
+# mnirs 0.1.6
+
+* Create internal numeric vector helper functions
+    * `signif_whole()` applies sig-fig rounding to decimal places, or whole values.
+    * `signif_trailing()` applies decimal or sig-fig rounding and convers to character strings with trailing zeroes, for display.
+    * `signif_pvalue()` applies formatting for p-value display as character strings or significance symbols.
+    * `seq_range()` creates a numeric sequence spanning the range of an ipnut vector.
+    * `wrap()` rotates vector elements from head to tail (or tail to head) by position.
+* Add statement about use of generative AI codebots to README.Rmd
+
 # mnirs 0.1.5
 
 * Remove redundant validation checks in `replace_mnirs()`.